string: Support targets with 16-bit ints in optimized code
The optimized string code is only used for 32-bit aligned data, but
the alignment checks were using 'long' instead of 'uintptr_t', causing
problems on targets with 16-bit pointers. Also, some computations need
explicit uint32_t casts to avoid truncation under the standard integer
promotion rules.

+++ newlib/libc/string/memset.c
@@ -46,7 +46,7 @@ QUICKREF
 #include "local.h"
 
 #define LBLOCKSIZE (sizeof(long))
-#define UNALIGNED(X)   ((long)X & (LBLOCKSIZE - 1))
+#define UNALIGNED(X)   ((uintptr_t)X & (LBLOCKSIZE - 1))
 #define TOO_SMALL(LEN) ((LEN) < LBLOCKSIZE)
 
 #undef memset
